requirements.txt
tensorflow>=1.15.0
numpy
Pillow